requirements.txt
APScheduler==3.5.3
dataclasses==0.6
Django==2.1.7
newsapi-python==0.2.3
pytz==2018.9
requests==2.17.1
six==1.12.0
sqlparse==0.3.0
tzlocal==1.5.1
urllib3==1.21.1
